ObserveDesingPattern
In this example, the Player class implements the IPlayer interface. Observer class Referee implements the IReferee interface. The player includes methods for registering, removing and updating watchers. The player's movement is calculated and the Direction, Angle and Force values are updated. These values are taken by the referee and printed on the screen.